allow selecting non-wiki-eligible objects
This change allows you to select any object in the QBE tree. For example, this is handy if you want to use QBE to Save off tiles for some objects that aren't wiki eligible. (You still can't upload anything that isn't wiki eligible.)

I also updated the styles in the tree a bit to make it clearer which things are wiki objects and which aren't.