fix regex (Graftek page issues)
The regex now REQUIRES the starting and ending '<!-- START QBE ... -->' and '<!-- END QBE -->' tags rather than making them optional. The previously non-required nature of these tags caused TEMPLATE_RE to match to a further part of the page that had an unrelated template that ended with '}}' on a new line. As a result, QBE deleted much of the page.